+++ internal/service/graph/scoring.go
@@ -14,7 +14,7 @@
 //  limitations under the License.
 //
 
-package kg
+package graph
 
 import (
 	"bytes"
