+++ internal/service/graph/search_test.go
@@ -14,7 +14,7 @@
 //  limitations under the License.
 //
 
-package kg
+package graph
 
 import (
 	"context"
