+++ internal/service/graph/testutil_test.go
@@ -1,3 +1,3 @@
-package kg
+package graph
 
 func strPtr(s string) *string { return &s }
